Fix: ã ì코드리뷰  피드백 적용

+++ backend/src/main/java/com/ssafy/yesrae/domain/article/repository/ArticleRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.yesrae.domain.article.repository;
 import com.ssafy.yesrae.domain.article.entity.ArticleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<ArticleEntity, Integer> {
+public interface ArticleRepository extends JpaRepository<ArticleEntity, Long> {
 }
